Implement full support for IPv6 in network library
Implements proper support for IPv6 addresses (this does not drop support for IPv4; it remains compatible!).

Address::setFromStr() is based on 2021 work I did on CryEngine (when upgrading the networking stack and all related tools to support IPv6 ontop of IPv4), and on several tools I've made for a heavily modified version of Valve Source. It has so far worked without issues for millions of users.

This patch also fixes a minor issue for windows where realloc might fail, but the old buffer was never free'd in that rare case.

Implementation (and the adaption of setFromStr) is currently untested!