Fix possible deadlock and race in getFFTBuffer() and pushFFT()
Always lock the mutex first before testing pointers, because it can become null during the lock if we test first resulting in a crash. It didn't incur a performance regression at all during my profiling tests. Also release mtx if rawFFTs is null (and don't acquire it if it is null at pushFFT).